requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==42.0.2
Django==4.2.3
django-environ==0.11.2
django-rest-knox==4.2.0
djangorestframework==3.14.0
pillow==10.2.0
psycopg==3.1.17
psycopg-binary==3.1.17
psycopg2-binary==2.9.6
pycparser==2.21
pytz==2023.4
sqlparse==0.4.4
typing_extensions==4.9.0